+++ heatmap-card.js
@@ -4079,7 +4079,7 @@ window.customCards.push({
     }
 });
 console.info(
-    "%c HEATMAP-CARD %c 2026.9.2-beta.1 ",
+    "%c HEATMAP-CARD %c 2026.9.2-beta.2 ",
     "color: black; background: #F2720C; font-weight: 600;",
     "color: black; background: #00a5c9; font-weight: 600;"
 );